requirements.txt
tensorflow>=2.13.0
albumentations>=1.3.0
opencv-python
matplotlib
seaborn
scikit-learn
pyyaml
lime
pytest
tqdm
Pillow